REF: Moves the activate user validation to a middleware
