+++ src/router/public/index.js
@@ -5,29 +5,13 @@ const router = express.Router()
 
 const reenableRouter = require('./reenableRoutes')
 
-const routerLogger = require('../logger')
 const pushpin = require('../../controllers/pushpin')
 const userController = require('../../controllers/user')
 const validateBody = require('../../middlewares/validateBody')
 const validateLoginBody = require('../../middlewares/validateLoginBody')
-const {
-  shouldActivateUser,
-  validateActivationToken
-} = require('../../validators/validateActivateUser')
+const validateActivation = require('../../middlewares/validateActivation')
 
-const {
-  MESSAGES,
-  COOKIE_KEY,
-  ERROR_HAS_OCCURRED
-} = require('../../constants')
-
-const logger = routerLogger.child({ module: 'public' })
-
-const errorResponse = ({ res, error }) => {
-  const { message, code, statusCode } = error
-
-  res.status(statusCode ?? 400).json({ message, code })
-}
+const { COOKIE_KEY } = require('../../constants')
 
 router.use(reenableRouter)
 
@@ -49,26 +33,16 @@ router.post('/login', validateLoginBody, async (req, res) => {
   userController.login(user, channel)
 })
 
-router.get('/:id/activate/:token', async (req, res) => {
-  const { params: { id, token } } = req
+router.get('/:id/activate/:token', validateActivation, async (req, res) => {
+  const { locals: { user, channel } } = res
 
-  try {
-    const user = await shouldActivateUser(id)
-    const decoded = await validateActivationToken(token)
-    const channel = decoded.channel.replace('rt:', '')
+  pushpin.sign.response({
+    res,
+    channel,
+    message: `Received activation token for user id: ${id}`
+  })
 
-    pushpin.sign.response({
-      res,
-      channel,
-      message: `Received activation token for user id: ${id}`
-    })
-
-    userController.activate({ user, channel })
-  } catch(error) {
-    logger.error({ error }, `${MESSAGES[ERROR_HAS_OCCURRED]} activate user`)
-
-    errorResponse({ res, error })
-  }
+  userController.activate({ user, channel })
 })
 
 module.exports = router
